fix(configuration): ensure proper font handling
kmscon has an issue when parsing the font spec from the configuration
file.  Fonts must be a list of comma-separated values; but if a single
font is present, unless a trailing comma is added, the font won't be
correctly set up.

After stylix started handling the font for kmscon, this broke my setup.
To fix it, I just left stylix keep its notion of monospace font, but I
forced my font spec to be at the head of the list.